pyproject.toml
[tool.poetry]
name = "messager"
version = "0.1.0"
description = ""
authors = ["Vadim <skolkovoinovations@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.2"
oauthlib = "^3.2.2"
pyjwt = "^2.8.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
